# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para sistema de gestión de biblioteca",
        "title": "Biblioteca API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/loans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of all loans in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get all loans (Admin only)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Loan status (active, returned, overdue)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loan"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/loans": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Request a book loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Create loan",
                "parameters": [
                    {
                        "description": "Loan request",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Loan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/loans/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get loan history for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get loan history",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LoanHistory"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/loans/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get current loans for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get my loans",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LoanResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/loans/{id}/return": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return a borrowed book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Return loan",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Book": {
            "type": "object",
            "required": [
                "author",
                "isbn",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "available_exemplars": {
                    "type": "integer"
                },
                "cover_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "exemplars": {
                    "description": "Relaciones",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Exemplar"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "isbn": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer",
                    "minimum": 1
                },
                "recommendations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Recommendation"
                    }
                },
                "title": {
                    "type": "string"
                },
                "total_exemplars": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Exemplar": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "acquisition_date": {
                    "type": "string"
                },
                "book": {
                    "description": "Relaciones",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Book"
                        }
                    ]
                },
                "book_id": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                },
                "loans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Loan"
                    }
                },
                "observations": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Fine": {
            "type": "object",
            "properties": {
                "accumulated_days": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relaciones",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.FineHistory": {
            "type": "object",
            "properties": {
                "accumulated_days": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "start_date": {
                    "type": "string"
                },
                "total_penalty_days": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relaciones",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Loan": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "exemplar": {
                    "$ref": "#/definitions/models.Exemplar"
                },
                "exemplar_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "loan_date": {
                    "type": "string"
                },
                "returned_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relaciones",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.LoanHistory": {
            "type": "object",
            "properties": {
                "book": {
                    "$ref": "#/definitions/models.Book"
                },
                "book_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "exemplar": {
                    "$ref": "#/definitions/models.Exemplar"
                },
                "exemplar_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "loan_date": {
                    "type": "string"
                },
                "overdue_days": {
                    "type": "integer"
                },
                "returned_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relaciones",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                },
                "was_overdue": {
                    "type": "boolean"
                }
            }
        },
        "models.LoanRequest": {
            "type": "object",
            "required": [
                "exemplar_id"
            ],
            "properties": {
                "exemplar_id": {
                    "type": "integer"
                }
            }
        },
        "models.LoanResponse": {
            "type": "object",
            "properties": {
                "book_author": {
                    "type": "string"
                },
                "book_title": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "exemplar_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_overdue": {
                    "type": "boolean"
                },
                "loan_date": {
                    "type": "string"
                },
                "overdue_days": {
                    "type": "integer"
                },
                "returned_at": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "last_name",
                "login",
                "name",
                "user_type"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "department_name": {
                    "description": "Solo profesores",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fine": {
                    "$ref": "#/definitions/models.Fine"
                },
                "fine_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FineHistory"
                    }
                },
                "floor": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "loan_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.LoanHistory"
                    }
                },
                "loans": {
                    "description": "Relaciones",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Loan"
                    }
                },
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "parents_phone": {
                    "description": "Campos específicos por tipo",
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.UserStatus"
                },
                "street": {
                    "description": "Dirección",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_type": {
                    "$ref": "#/definitions/models.UserType"
                }
            }
        },
        "models.UserStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "DEBTOR",
                "FINED"
            ],
            "x-enum-varnames": [
                "ACTIVE",
                "DEBTOR",
                "FINED"
            ]
        },
        "models.UserType": {
            "type": "string",
            "enum": [
                "STUDENT",
                "TEACHER",
                "ADMIN"
            ],
            "x-enum-varnames": [
                "STUDENT",
                "TEACHER",
                "ADMIN"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}